+++ projects/ngx-ui-tour-ionic/src/public_api.ts
@@ -4,4 +4,5 @@ export * from './lib/tour-ion-popover.module';
 export * from './lib/tour-anchor.directive';
 
 export {IonStepOption as IStepOption, IonPopoverPlacement} from './lib/step-option.interface';
+export {StepDimensions} from 'ngx-ui-tour-core';
 export {IonTourService as TourService} from './lib/ion-tour.service';