+++ projects/ngx-ui-tour-ng-bootstrap/src/public_api.ts
@@ -3,4 +3,5 @@ export {TourStepTemplateComponent} from './lib/tour-step-template.component';
 export {TourNgBootstrapModule} from './lib/ng-bootstrap.module';
 
 export {INgbStepOption as IStepOption} from './lib/step-option.interface';
+export {StepDimensions} from 'ngx-ui-tour-core';
 export {NgbTourService as TourService} from './lib/ng-bootstrap-tour.service';