+++ projects/ngx-ui-tour-ngx-bootstrap/src/public_api.ts
@@ -3,4 +3,5 @@ export {TourAnchorNgxBootstrapDirective, TourAnchorNgxBootstrapPopoverDirective}
 export {TourNgxBootstrapModule} from './lib/ngx-bootstrap.module';
 
 export {INgxbStepOption as IStepOption} from './lib/step-option.interface';
+export {StepDimensions} from 'ngx-ui-tour-core';
 export {NgxbTourService as TourService} from './lib/ngx-bootstrap-tour.service';