+++ projects/ngx-ui-tour-tui-dropdown/src/public_api.ts
@@ -3,4 +3,5 @@ export {TourAnchorTuiDropdownDirective} from './lib/tour-anchor.directive';
 export {TourTuiDropdownModule} from './lib/tour-tui-dropdown.module';
 
 export {ITuiDdStepOption as IStepOption, TuiDdPlacement} from './lib/step-option.interface';
+export {StepDimensions} from 'ngx-ui-tour-core';
 export {TourTuiDropdownService as TourService} from './lib/tour-tui-dropdown.service';