+++ projects/ngx-ui-tour-console/CHANGELOG.md
@@ -1,3 +1,19 @@
+<a name="10.5.1"></a>
+
+# [10.5.1](https://github.com/hakimio/ngx-ui-tour) (2023-03-14)
+
+### Fixes
+- Fix default step max-width.
+
+<a name="10.5.0"></a>
+
+# [10.5.0](https://github.com/hakimio/ngx-ui-tour) (2023-03-13)
+
+### Features
+- New `allowUserInitiatedNavigation` config. Enable to keep the tour open while the user navigates to a different page.
+  Can cause the tour to get "stuck" if active anchor element is removed from the DOM after the navigation.
+  The option is not related to navigation initiated by the UI tour and should not be enabled in most cases.
+
 <a name="10.4.0"></a>
 
 # [10.4.0](https://github.com/hakimio/ngx-ui-tour) (2023-02-14)